ensure remote builder only once config is valid
