+++ docker/build.go
@@ -3,21 +3,27 @@ package docker
 import (
 	"errors"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"net/http"
+	"net/url"
 	"os"
 	"path"
 	"strings"
 	"time"
 
+	"github.com/dustin/go-humanize"
+	"github.com/jpillora/backoff"
+	"github.com/superfly/flyctl/api"
 	"github.com/superfly/flyctl/builtinsupport"
 	"github.com/superfly/flyctl/cmdctx"
+	"golang.org/x/net/context"
+	"golang.org/x/net/http/httpproxy"
 
 	"github.com/briandowns/spinner"
 	"github.com/buildpacks/pack"
 	"github.com/docker/docker/builder/dockerignore"
+	"github.com/docker/docker/client"
 	"github.com/docker/docker/pkg/fileutils"
-	"github.com/superfly/flyctl/api"
 	"github.com/superfly/flyctl/flyctl"
 	"github.com/superfly/flyctl/helpers"
 	"github.com/superfly/flyctl/terminal"
@@ -32,35 +38,149 @@ var ErrDockerDaemon = errors.New("Docker daemon must be running to perform this 
 // ErrNoBuildpackBuilder - Unable to find Buildpack builder
 var ErrNoBuildpackBuilder = errors.New("No buildpack builder")
 
-// ResolveDockerfile - Resolve the location of the dockerfile, allowing for upper and lowercase naming
-func ResolveDockerfile(cwd string) string {
-	dockerfilePath := path.Join(cwd, "Dockerfile")
-	if helpers.FileExists(dockerfilePath) {
-		return dockerfilePath
-	}
-	dockerfilePath = path.Join(cwd, "dockerfile")
-	if helpers.FileExists(dockerfilePath) {
-		return dockerfilePath
-	}
-	return ""
+type BuildOperation struct {
+	ctx                  context.Context
+	apiClient            *api.Client
+	dockerClient         *DockerClient
+	localDockerAvailable bool
+	out                  io.Writer
+	appName              string
+	appConfig            *flyctl.AppConfig
+	imageTag             string
+	remoteOnly           bool
+	localOnly            bool
 }
 
-// Image - A type to hold information about a Docker image, including ID, Tag and Size
-type Image struct {
-	ID   string
-	Tag  string
-	Size int64
+func NewBuildOperation(ctx context.Context, cmdCtx *cmdctx.CmdContext) (*BuildOperation, error) {
+	remoteOnly := cmdCtx.Config.GetBool("remote-only")
+	localOnly := cmdCtx.Config.GetBool("local-only")
+
+	if localOnly && remoteOnly {
+		return nil, fmt.Errorf("Both --local-only and --remote-only are set - select only one")
+	}
+
+	imageLabel, _ := cmdCtx.Config.GetString("image-label")
+
+	dockerClient, err := NewDockerClient()
+	if err != nil {
+		return nil, err
+	}
+
+	op := &BuildOperation{
+		ctx:          ctx,
+		dockerClient: dockerClient,
+		apiClient:    cmdCtx.Client.API(),
+		out:          cmdCtx.Out,
+		appName:      cmdCtx.AppName,
+		appConfig:    cmdCtx.AppConfig,
+		imageTag:     newDeploymentTag(cmdCtx.AppName, imageLabel),
+		localOnly:    localOnly,
+		remoteOnly:   remoteOnly,
+	}
+
+	if err := op.dockerClient.Check(ctx); err == nil {
+		op.localDockerAvailable = true
+	} else {
+		terminal.Debugf("Error pinging local docker: %s\n", err)
+	}
+
+	if remoteOnly {
+		terminal.Info("Remote only, hooking you up with a remote Docker builder...")
+		if err := setRemoteBuilder(ctx, cmdCtx, dockerClient); err != nil {
+			return nil, err
+		}
+	} else if err := op.dockerClient.Check(ctx); err != nil {
+		if localOnly {
+			return nil, fmt.Errorf("Local docker unavailable and --local-only was passed, cannot proceed.")
+		}
+		terminal.Info("Local docker unavailable, hooking you up with a remote Docker builder...")
+		if err := setRemoteBuilder(ctx, cmdCtx, dockerClient); err != nil {
+			return nil, err
+		}
+	}
+
+	return op, nil
+}
+
+func (op *BuildOperation) LocalDockerAvailable() bool {
+	return op.localDockerAvailable
+}
+
+func (op *BuildOperation) LocalOnly() bool {
+	return op.localOnly
+}
+
+func (op *BuildOperation) RemoteOnly() bool {
+	return op.remoteOnly
+}
+
+func (op *BuildOperation) ResolveImageLocally(ctx context.Context, cmdCtx *cmdctx.CmdContext, imageRef string) (*Image, error) {
+	cmdCtx.Status("deploy", "Resolving image")
+
+	if !op.LocalDockerAvailable() || op.RemoteOnly() {
+		return nil, nil
+	}
+
+	imgSummary, err := op.dockerClient.findImage(ctx, imageRef)
+	if err != nil {
+		return nil, err
+	}
+
+	if imgSummary == nil {
+		return nil, nil
+	}
+
+	cmdCtx.Statusf("deploy", cmdctx.SINFO, "Image ID: %+v\n", imgSummary.ID)
+	cmdCtx.Statusf("deploy", cmdctx.SINFO, "Image size: %s\n", humanize.Bytes(uint64(imgSummary.Size)))
+
+	cmdCtx.Status("deploy", cmdctx.SDONE, "Image resolving done")
+
+	cmdCtx.Status("deploy", cmdctx.SBEGIN, "Creating deployment tag")
+	if err := op.dockerClient.TagImage(op.ctx, imgSummary.ID, op.imageTag); err != nil {
+		return nil, err
+	}
+	cmdCtx.Status("deploy", cmdctx.SINFO, "-->", op.imageTag)
+
+	image := &Image{
+		ID:   imgSummary.ID,
+		Size: imgSummary.Size,
+		Tag:  op.imageTag,
+	}
+
+	err = op.PushImage(*image)
+
+	if err != nil {
+		return nil, err
+	}
+
+	return image, nil
+}
+
+func (op *BuildOperation) pushImage(imageTag string) error {
+
+	if imageTag == "" {
+		return errors.New("invalid image reference")
+	}
+
+	if err := op.dockerClient.PushImage(op.ctx, imageTag, op.out); err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func (op *BuildOperation) CleanDeploymentTags() {
+	err := op.dockerClient.DeleteDeploymentImages(op.ctx, op.imageTag)
+	if err != nil {
+		terminal.Debugf("Error cleaning deployment tags: %s", err)
+	}
 }
 
 // BuildWithDocker - Run a Docker Build operation reporting back via the command context
-func (op *DeployOperation) BuildWithDocker(cmdCtx *cmdctx.CmdContext, dockerfilePath string, buildArgs map[string]string) (*Image, error) {
+func (op *BuildOperation) BuildWithDocker(cmdCtx *cmdctx.CmdContext, dockerfilePath string, buildArgs map[string]string) (*Image, error) {
 	spinning := cmdCtx.OutputJSON()
 	cwd := cmdCtx.WorkingDir
 	appConfig := cmdCtx.AppConfig
-
-	if !op.DockerAvailable() {
-		return nil, ErrDockerDaemon
-	}
 
 	if dockerfilePath == "" {
 		dockerfilePath = ResolveDockerfile(cwd)
@@ -151,8 +271,8 @@ func (op *DeployOperation) BuildWithDocker(cmdCtx *cmdctx.CmdContext, dockerfile
 	return image, nil
 }
 
-func initPackClient() pack.Client {
-	client, err := pack.NewClient()
+func (op *BuildOperation) initPackClient() pack.Client {
+	client, err := pack.NewClient(pack.WithDockerClient(op.dockerClient.docker))
 	if err != nil {
 		panic(err)
 	}
@@ -160,19 +280,15 @@ func initPackClient() pack.Client {
 }
 
 // BuildWithPack - Perform a Docker build using a Buildpack (buildpack.io)
-func (op *DeployOperation) BuildWithPack(cmdCtx *cmdctx.CmdContext, buildArgs map[string]string) (*Image, error) {
+func (op *BuildOperation) BuildWithPack(cmdCtx *cmdctx.CmdContext, buildArgs map[string]string) (*Image, error) {
 	cwd := cmdCtx.WorkingDir
 	appConfig := cmdCtx.AppConfig
-
-	if !op.DockerAvailable() {
-		return nil, ErrDockerDaemon
-	}
 
 	if appConfig.Build == nil || appConfig.Build.Builder == "" {
 		return nil, ErrNoBuildpackBuilder
 	}
 
-	c := initPackClient()
+	c := op.initPackClient()
 
 	env := map[string]string{}
 
@@ -216,91 +332,28 @@ func (op *DeployOperation) BuildWithPack(cmdCtx *cmdctx.CmdContext, buildArgs ma
 }
 
 // PushImage - Push the Image (where?)
-func (op *DeployOperation) PushImage(image Image) error {
+func (op *BuildOperation) PushImage(image Image) error {
 	return op.pushImage(image.Tag)
 }
 
-// StartRemoteBuild - Start a remote build and track its progress
-func (op *DeployOperation) StartRemoteBuild(cwd string, appConfig *flyctl.AppConfig, dockerfilePath string, buildArgs map[string]string) (*api.Build, error) {
-	buildContext, err := newBuildContext()
-	if err != nil {
-		return nil, err
+// ResolveDockerfile - Resolve the location of the dockerfile, allowing for upper and lowercase naming
+func ResolveDockerfile(cwd string) string {
+	dockerfilePath := path.Join(cwd, "Dockerfile")
+	if helpers.FileExists(dockerfilePath) {
+		return dockerfilePath
 	}
-	defer buildContext.Close()
-
-	s := spinner.New(spinner.CharSets[11], 100*time.Millisecond)
-	s.Writer = os.Stderr
-	s.Prefix = "Creating build context... "
-	s.Start()
-
-	excludes, err := readGitignore(cwd)
-	if err != nil {
-		return nil, err
+	dockerfilePath = path.Join(cwd, "dockerfile")
+	if helpers.FileExists(dockerfilePath) {
+		return dockerfilePath
 	}
+	return ""
+}
 
-	if err := buildContext.AddSource(cwd, excludes); err != nil {
-		return nil, err
-	}
-
-	if dockerfilePath != "" {
-		dockerfile, err := os.Open(dockerfilePath)
-		if err != nil {
-			return nil, err
-		}
-		defer dockerfile.Close()
-		if err := buildContext.AddFile("Dockerfile", dockerfile); err != nil {
-			return nil, err
-		}
-	}
-
-	archive, err := buildContext.Archive()
-	if err != nil {
-		return nil, err
-	}
-	defer archive.Close()
-	s.Stop()
-
-	s.Prefix = "Submitting build..."
-
-	uploadFileName := fmt.Sprintf("source-%d.tar.gz", time.Now().Unix())
-	getURL, putURL, err := op.apiClient.CreateSignedUrls(op.AppName(), uploadFileName)
-	if err != nil {
-		return nil, err
-	}
-
-	req, err := http.NewRequest("PUT", putURL, archive.File)
-	if err != nil {
-		return nil, err
-	}
-	req.ContentLength = archive.Size
-
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	if resp.StatusCode >= 300 {
-		body, _ := ioutil.ReadAll(resp.Body)
-		return nil, fmt.Errorf("Error submitting build: %s", body)
-	}
-
-	input := api.StartBuildInput{
-		AppID:      op.AppName(),
-		SourceURL:  getURL,
-		SourceType: "targz",
-		BuildType:  api.StringPointer("flyctl_v1"),
-	}
-
-	for name, val := range buildArgs {
-		input.BuildArgs = append(input.BuildArgs, api.BuildArgInput{Name: name, Value: val})
-	}
-
-	build, err := op.apiClient.StartBuild(input)
-	if err != nil {
-		return nil, err
-	}
-	s.Stop()
-
-	return build, nil
+// Image - A type to hold information about a Docker image, including ID, Tag and Size
+type Image struct {
+	ID   string
+	Tag  string
+	Size int64
 }
 
 func normalizeBuildArgs(appConfig *flyctl.AppConfig, extra map[string]string) map[string]*string {
@@ -340,23 +393,6 @@ func readDockerignore(workingDir string) ([]string, error) {
 	return excludes, err
 }
 
-func readGitignore(workingDir string) ([]string, error) {
-	file, err := os.Open(path.Join(workingDir, ".gitignore"))
-	if os.IsNotExist(err) {
-		return []string{}, nil
-	} else if err != nil {
-		terminal.Warn("Error reading gitignore", err)
-		return []string{}, nil
-	}
-
-	excludes, err := dockerignore.ReadAll(file)
-	if err == nil {
-		excludes = trimExcludes(excludes)
-	}
-
-	return excludes, err
-}
-
 func trimExcludes(excludes []string) []string {
 	if match, _ := fileutils.Matches(".dockerignore", excludes); match {
 		excludes = append(excludes, "!.dockerignore")
@@ -367,4 +403,88 @@ func trimExcludes(excludes []string) []string {
 	}
 
 	return excludes
+}
+
+func setRemoteBuilder(ctx context.Context, cmdCtx *cmdctx.CmdContext, dockerClient *DockerClient) error {
+	rawURL, release, err := cmdCtx.Client.API().EnsureRemoteBuilder(cmdCtx.AppName)
+	if err != nil {
+		return fmt.Errorf("could not create remote builder: %v", err)
+	}
+
+	terminal.Debugf("Remote Docker builder URL: %s\n", rawURL)
+	terminal.Debugf("Remote Docker builder release: %+v\n", release)
+
+	dockerClient.docker, err = newDockerClient(client.WithHost(fmt.Sprintf("tcp://%s", cmdCtx.AppName)))
+	if err != nil {
+		return fmt.Errorf("error resetting docker client to use remote builder config: %v", err)
+	}
+
+	dockerTransport, ok := dockerClient.docker.HTTPClient().Transport.(*http.Transport)
+	if !ok {
+		return fmt.Errorf("Docker client transport was not an HTTP transport, don't know what to do with that")
+	}
+	builderURL, err := url.Parse(rawURL)
+	if err != nil {
+		return fmt.Errorf("Could not parse builder url '%s': %v", rawURL, err)
+	}
+
+	builderURL.User = url.UserPassword(cmdCtx.AppName, flyctl.GetAPIToken())
+
+	proxyCfg := &httpproxy.Config{HTTPProxy: builderURL.String()}
+	dockerTransport.Proxy = func(req *http.Request) (*url.URL, error) {
+		return proxyCfg.ProxyFunc()(req.URL)
+	}
+
+	deadline := time.After(5 * time.Minute)
+
+	terminal.Info("Waiting for remote builder to become available...")
+
+	b := &backoff.Backoff{
+		//These are the defaults
+		Min:    200 * time.Millisecond,
+		Max:    2 * time.Second,
+		Factor: 1.2,
+		Jitter: true,
+	}
+
+	consecutiveSuccesses := 0
+
+OUTER:
+	for {
+		checkErr := make(chan error, 1)
+
+		go func() {
+			checkErr <- dockerClient.Check(ctx)
+		}()
+
+		select {
+		case err := <-checkErr:
+			if err == nil {
+				if consecutiveSuccesses == 0 {
+					// reset on the first success in a row so the next checks are a bit spaced out
+					b.Reset()
+				}
+				consecutiveSuccesses++
+				if consecutiveSuccesses >= 3 {
+					terminal.Info("Remote builder is ready to build!")
+					break OUTER
+				}
+				dur := b.Duration()
+				terminal.Debugf("Remote builder available, but pinging again in %s to be sure\n", dur)
+				time.Sleep(dur)
+			} else {
+				consecutiveSuccesses = 0
+				dur := b.Duration()
+				terminal.Debugf("Remote builder unavailable, retrying in %s (err: %v)\n", dur, err)
+				time.Sleep(dur)
+			}
+		case <-deadline:
+			return fmt.Errorf("Could not ping remote builder within 5 minutes, aborting.")
+		case <-ctx.Done():
+			terminal.Warn("Canceled")
+			break OUTER
+		}
+	}
+
+	return nil
 }
